+++ src/main/java/core/jdbc/JdbcTemplate.java
@@ -1,6 +1,4 @@
-package jwp.dao;
-
-import core.jdbc.ConnectionManager;
+package core.jdbc;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
