+++ src/main/java/core/jdbc/PreparedStatementSetter.java
@@ -1,4 +1,4 @@
-package jwp.dao;
+package core.jdbc;
 
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
