+++ src/main/java/core/jdbc/RowMapper.java
@@ -1,4 +1,4 @@
-package jwp.dao;
+package core.jdbc;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
